+++ packages/snow_rendering/lib/snow_rendering.dart
@@ -24,5 +24,6 @@ export 'src/assessment/driving_condition_assessment.dart';
 export 'src/data_budget.dart';
 export 'src/models/precipitation_config.dart';
 export 'src/models/recommended_response.dart';
+export 'src/models/road_surface_announcement.dart';
 export 'src/models/road_surface_state.dart';
 export 'src/models/visibility_degradation.dart';